extract routing server data creation routines into standalone module
